refactor(server): extract advanceWeekService from gameLoop route
Phase 2 PR-3. Moves the POST /api/advance-week engine orchestration out of the
fat gameLoop.ts handler into server/services/advanceWeekService.ts, following
the gameCreationService / saveService convention (class + exported singleton,
deps defaulted to module singletons).

Verbatim move (D6: transactionality out of scope):
 - Both transactions move unchanged, split at the SAME boundary (engine
   execution in TX 1, persistence + debug read-back + response assembly in
   TX 2). Same operations, same order in each.
 - Every console.log/console.error, the GameEngine construction, the
   campaign-completed early return, the action formatting, and the response
   envelope ({ gameState, summary, campaignResults, debug }) are byte-identical.

Route thins to: middleware chain (requireClerkUser + requireGameOwner)
unchanged -> validateRequest(AdvanceWeekRequest) -> service -> res.json. All
HTTP status mapping stays in the route (ZodError -> 400 VALIDATION_ERROR,
everything else -> 500 ADVANCE_WEEK_ERROR) with byte-identical bodies. The
service throws plain Errors, mirroring the handler's inline throws.

gameLoop.ts 407 -> 75 lines; advanceWeekService.ts 385 lines.

Gates green: npm run check; golden-master-advance-week, sweep-migrations +
advance-week characterizations, game-engine-advance-week-integration,
game-spine-advance-save-restore, route-manifest. Golden master byte-identical,
manifest unchanged.
